Update input_parser.py
Here's the parser with some modifications. You can see its usage in the file substrate_slider.py
There are some possibly problematic naming, such as "parm", "parms" for the word "parameter(s)", so we may discuss about it. 
All the files that are pull-requested are working together, so you can run them.