Update and rename parameters.py to substrate_slider.py
By calling the parser two times for different classes, different blocks of the input file are parsed. 
With the parsed parameters, two classes are created.
Then, these classes can be imported into a different file. I created the file try_classes.py for trying the code.
I know I was responsible only for the class Substrate, not Slider, but I had to be sure that the parser was working with different classes. That's why I added the class Slider. The parameters for Slider are just arbitrary now, and should be changed.